pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]


[tool.black]
extend-exclude = '''
(
    .*\.local.*
    | .*\.idea.*
    | .*\.vscode.*
)
'''
include = '^.+\.pyi?$'
line-length = 79
target-version = ["py311"]


[tool.django-stubs]
django_settings_module = "project.settings"


[tool.isort]
force_single_line = true
line_length = 79
multi_line_output = 3
profile = "black"
skip = [
    ".local",
]


[tool.mypy]
cache_dir = ".local/mypy/"
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
exclude=[".local/*",]
explicit_package_bases = true
files = ["."]
no_implicit_optional = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
    "pydantic.mypy",
]
pretty = true
python_version = "3.11"
show_error_context = true
sqlite_cache = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true


[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true


[tool.pytest.ini_options]
cache_dir = ".local/pytest/"
console_output_style = "count"
testpaths = [
    "tests",
]


[tool.poetry]
name = "example-ca"
version = "2023.11.15"
description = "Example: Clean Architecture"
authors = ["Alexander Sidorov <alexander@sidorov.dev>"]
readme = "README.md"


[tool.poetry.dependencies]
attrs = "23.1.0"
cython = "3.0.5"
devtools = {extras = ["pygments"], version = "0.12.2"}
dj-database-url = "2.1.0"
django = "4.2.7"
djangorestframework = "3.14.0"
gunicorn = "21.2.0"
httpx = "0.25.1"
orjson = "3.9.10"
pendulum = "2.1.2"
psycopg2-binary = "2.9.9"
pydantic-settings = "2.1.0"
python = "3.11.5"
sqlalchemy = "2.0.23"
tenacity = "8.2.3"


[tool.poetry.group.dev.dependencies]
autoflake = "2.2.1"
black = "23.11.0"
django-stubs = "4.2.6"
djangorestframework-stubs = "3.14.4"
flake8 = "6.1.0"
flake8-annotations-complexity = "0.0.8"
flake8-bandit = "4.1.1"
flake8-black = "0.3.6"
flake8-bugbear = "23.9.16"
flake8-builtins = "2.2.0"
flake8-cognitive-complexity = "0.1.0"
flake8-comprehensions = "3.14.0"
flake8-debugger = "4.1.2"
flake8-executable = "2.1.3"
flake8-expression-complexity = "0.0.11"
flake8-fixme = "1.1.1"
flake8-isort = "6.1.1"
flake8-multiline-containers = "0.0.19"
flake8-mutable = "1.2.0"
flake8-pep3101 = "2.1.0"
flake8-print = "5.0.0"
flake8-printf-formatting = "1.1.2"
flake8-simplify = "0.21.0"
flake8-variables-names = "0.0.6"
isort = "5.12.0"
mypy = "1.7.0"
pep8-naming = "0.13.3"
pytest = "7.4.3"
types-psycopg2 = "2.9.21.16"
